refactor(write): split postpone fixed-bucket modules
