+++ crates/paimon/src/table/mod.rs
@@ -72,8 +72,11 @@ mod pk_vector_indexed_split_read;
 mod pk_vector_orchestrator;
 mod pk_vector_position_read;
 mod pk_vector_scan;
-mod postpone_batch_table_write;
+mod postpone_bucket_plan;
 mod postpone_file_writer;
+mod postpone_fixed_bucket_router;
+mod postpone_fixed_bucket_write;
+mod postpone_fixed_bucket_write_builder;
 mod prepared_files;
 mod read_builder;
 pub mod referenced_files;
@@ -121,11 +124,12 @@ pub use incremental_scan::{
 };
 pub use lumina_index_build_builder::LuminaIndexBuildBuilder;
 pub use partition_stat::PartitionStat;
-pub use postpone_batch_table_write::{
-    PostponeBucketBatch, PostponeBucketPlan, PostponeFixedBucketRouter,
-    PostponeFixedBucketTableCommit, PostponeFixedBucketTableWrite, PostponeFixedBucketWriteBuilder,
-    POSTPONE_BUCKET_PLAN_TOTAL_BUCKETS_FIELD,
+pub use postpone_bucket_plan::{PostponeBucketPlan, POSTPONE_BUCKET_PLAN_TOTAL_BUCKETS_FIELD};
+pub use postpone_fixed_bucket_router::{PostponeBucketBatch, PostponeFixedBucketRouter};
+pub use postpone_fixed_bucket_write::{
+    PostponeFixedBucketTableCommit, PostponeFixedBucketTableWrite,
 };
+pub use postpone_fixed_bucket_write_builder::PostponeFixedBucketWriteBuilder;
 pub use read_builder::ReadBuilder;
 pub use rest_env::RESTEnv;
 pub use scan_trace::ScanTrace;
